+++ src/Products/urban/docgen/helper_view.py
@@ -4,12 +4,38 @@ from collective.documentgenerator.helper.archetypes import ATDocumentGenerationH
 from datetime import date as _date
 from Products.CMFPlone.i18nl10n import ulocalized_time
 from zope.i18n import translate
+from plone import api
 
 
 class UrbanDocGenerationHelperView(ATDocumentGenerationHelperView):
     """
     Urban implementation of document generation helper methods.
     """
+
+    def get_work_location(self, index):
+        """
+        return a dictionary containing all work locations informations
+        """
+        work_location = {
+                'bestAddressKey': '',
+                'streetCode': '',
+                'streetName': '',
+                'startDate': '',
+                'endDate': '',
+                'regionalRoad': '',
+                'number': '',
+        }
+        workLocation = self.context.getWorkLocations()[index]
+        catalog = api.portal.get_tool("uid_catalog")
+        street = catalog(UID=workLocation['street'])[0].getObject()
+        work_location['bestAddressKey'] = street.getBestAddressKey()
+        work_location['streetCode'] = street.getStreetCode()
+        work_location['streetName'] = street.getStreetName()
+        work_location['startDate'] = street.getStartDate()
+        work_location['endDate'] = street.getEndDate()
+        work_location['regionalRoad'] = street.getRegionalRoad()
+        work_location['number'] =  workLocation['number']
+        return work_location
 
     def formatDate(self, date=_date.today(), translatemonth=True, long_format=False):
         """